remove layout table v1 and rename v2 to v1
