+++ pkg/vm/engine/aoe/storage/db/gcreqs/droptbl.go
@@ -5,10 +5,9 @@ import (
 	"matrixone/pkg/vm/engine/aoe/storage/dbi"
 	"matrixone/pkg/vm/engine/aoe/storage/events/memdata"
 	"matrixone/pkg/vm/engine/aoe/storage/gc"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1"
 	mtif "matrixone/pkg/vm/engine/aoe/storage/memtable/base"
 	"matrixone/pkg/vm/engine/aoe/storage/ops"
-	// log "github.com/sirupsen/logrus"
 )
 
 type dropTblRequest struct {
