+++ pkg/vm/engine/aoe/storage/db/sched/scheduler.go
@@ -3,7 +3,7 @@ package sched
 import (
 	"matrixone/pkg/logutil"
 	e "matrixone/pkg/vm/engine/aoe/storage"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1"
 	md "matrixone/pkg/vm/engine/aoe/storage/metadata/v1"
 	"matrixone/pkg/vm/engine/aoe/storage/sched"
 	"sync"
