+++ pkg/vm/engine/aoe/storage/db/sched/upgradeseg.go
@@ -1,9 +1,8 @@
 package sched
 
 import (
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2/iface"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/iface"
 	"matrixone/pkg/vm/engine/aoe/storage/sched"
-	// log "github.com/sirupsen/logrus"
 )
 
 type upgradeSegEvent struct {
