+++ pkg/vm/engine/aoe/storage/events/memdata/droptbl.go
@@ -1,7 +1,7 @@
 package memdata
 
 import (
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2/iface"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/iface"
 	"matrixone/pkg/vm/engine/aoe/storage/sched"
 )
