+++ pkg/vm/engine/aoe/storage/layout/table/v1/blk.go
@@ -11,9 +11,9 @@ import (
 	"matrixone/pkg/vm/engine/aoe/storage/dbi"
 	"matrixone/pkg/vm/engine/aoe/storage/layout/base"
 	"matrixone/pkg/vm/engine/aoe/storage/layout/index"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2/col"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2/iface"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2/wrapper"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/col"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/iface"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/wrapper"
 	md "matrixone/pkg/vm/engine/aoe/storage/metadata/v1"
 	"matrixone/pkg/vm/process"
 	// log "github.com/sirupsen/logrus"
