+++ pkg/vm/engine/aoe/storage/layout/table/v1/col/blk.go
@@ -1,54 +1,53 @@
 package col
 
 import (
-	"io"
+	"matrixone/pkg/container/types"
+	ro "matrixone/pkg/container/vector"
 	"matrixone/pkg/vm/engine/aoe/storage/common"
+	"matrixone/pkg/vm/engine/aoe/storage/container/vector"
+	"matrixone/pkg/vm/engine/aoe/storage/dbi"
 	"matrixone/pkg/vm/engine/aoe/storage/layout/base"
 	"matrixone/pkg/vm/engine/aoe/storage/layout/index"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/iface"
 	md "matrixone/pkg/vm/engine/aoe/storage/metadata/v1"
+	"matrixone/pkg/vm/process"
 	"sync"
 	"sync/atomic"
-	// log "github.com/sirupsen/logrus"
 )
 
 type IColumnBlock interface {
-	io.Closer
-	GetNext() IColumnBlock
-	SetNext(next IColumnBlock)
-	GetID() common.ID
+	common.IRef
+	GetID() uint64
+	GetMeta() *md.Block
 	GetRowCount() uint64
-	InitScanCursor(cusor *ScanCursor) error
-	Append(part IColumnPart)
-	GetPartRoot() IColumnPart
-	GetBlockType() base.BlockType
+	RegisterPart(part IColumnPart)
+	GetType() base.BlockType
+	GetColType() types.Type
 	GetIndexHolder() *index.BlockHolder
 	GetColIdx() int
-	CloneWithUpgrade(IColumnSegment, *md.Block) IColumnBlock
-	EvalFilter(*index.FilterCtx) error
+	GetSegmentFile() base.ISegmentFile
+	CloneWithUpgrade(iface.IBlock) IColumnBlock
 	String() string
-	Ref() IColumnBlock
-	UnRef()
-	GetRefs() int64
+	Size() uint64
+	GetVector() vector.IVector
+	LoadVectorWrapper() (*vector.VectorWrapper, error)
+	ForceLoad(ref uint64, proc *process.Process) (*ro.Vector, error)
+	Prefetch() error
+	GetVectorReader() dbi.IVectorReader
 }
 
 type ColumnBlock struct {
 	sync.RWMutex
-	ID          common.ID
-	Next        IColumnBlock
-	Type        base.BlockType
+	common.RefHelper
 	ColIdx      int
-	Refs        int64
 	Meta        *md.Block
-	File        base.ISegmentFile
+	SegmentFile base.ISegmentFile
 	IndexHolder *index.BlockHolder
+	Type        base.BlockType
 }
 
-func (blk *ColumnBlock) EvalFilter(ctx *index.FilterCtx) error {
-	return blk.IndexHolder.EvalFilter(blk.ColIdx, ctx)
-}
-
-func (blk *ColumnBlock) GetRefs() int64 {
-	return atomic.LoadInt64(&blk.Refs)
+func (blk *ColumnBlock) GetSegmentFile() base.ISegmentFile {
+	return blk.SegmentFile
 }
 
 func (blk *ColumnBlock) GetIndexHolder() *index.BlockHolder {
@@ -59,9 +58,15 @@ func (blk *ColumnBlock) GetColIdx() int {
 	return blk.ColIdx
 }
 
-func (blk *ColumnBlock) GetBlockType() base.BlockType {
-	blk.RLock()
-	defer blk.RUnlock()
+func (blk *ColumnBlock) GetColType() types.Type {
+	return blk.Meta.Segment.Table.Schema.ColDefs[blk.ColIdx].Type
+}
+
+func (blk *ColumnBlock) GetMeta() *md.Block {
+	return blk.Meta
+}
+
+func (blk *ColumnBlock) GetType() base.BlockType {
 	return blk.Type
 }
 
@@ -69,25 +74,6 @@ func (blk *ColumnBlock) GetRowCount() uint64 {
 	return atomic.LoadUint64(&blk.Meta.Count)
 }
 
-func (blk *ColumnBlock) SetNext(next IColumnBlock) {
-	blk.Lock()
-	defer blk.Unlock()
-	if blk.Next != nil {
-		blk.Next.UnRef()
-	}
-	blk.Next = next
-}
-
-func (blk *ColumnBlock) GetNext() IColumnBlock {
-	blk.RLock()
-	if blk.Next != nil {
-		blk.Next.Ref()
-	}
-	r := blk.Next
-	blk.RUnlock()
-	return r
-}
-
-func (blk *ColumnBlock) GetID() common.ID {
-	return blk.ID
+func (blk *ColumnBlock) GetID() uint64 {
+	return blk.Meta.ID
 }
