+++ pkg/vm/engine/aoe/storage/layout/table/v1/data.go
@@ -4,88 +4,96 @@ import (
 	"errors"
 	"fmt"
 	"matrixone/pkg/container/types"
-	logutil2 "matrixone/pkg/logutil"
 	bmgrif "matrixone/pkg/vm/engine/aoe/storage/buffer/manager/iface"
 	"matrixone/pkg/vm/engine/aoe/storage/common"
 	"matrixone/pkg/vm/engine/aoe/storage/layout/base"
 	"matrixone/pkg/vm/engine/aoe/storage/layout/index"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/col"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/iface"
 	md "matrixone/pkg/vm/engine/aoe/storage/metadata/v1"
-	"runtime"
 	"sync"
+	"sync/atomic"
 )
 
-type ITableData interface {
-	sync.Locker
-	GetRowCount() uint64
-	GetID() uint64
-	GetCollumns() []col.IColumnData
-	GetCollumn(int) col.IColumnData
-	GetColTypeSize(idx int) uint64
-	GetColTypes() []types.Type
-	GetMTBufMgr() bmgrif.IBufferManager
-	GetSSTBufMgr() bmgrif.IBufferManager
-	GetSegmentCount() uint64
-	GetIndexHolder() *index.TableHolder
+var (
+	NotExistErr = errors.New("not exist error")
+)
 
-	UpgradeBlock(*md.Block) (blks []col.IColumnBlock)
-	UpgradeSegment(segID common.ID) (segs []col.IColumnSegment)
-	AppendColSegments(colSegs []col.IColumnSegment)
-}
-
-func NewTableData(fsMgr base.IManager, indexBufMgr, mtBufMgr, sstBufMgr bmgrif.IBufferManager, meta *md.Table) ITableData {
+func NewTableData(fsMgr base.IManager, indexBufMgr, mtBufMgr, sstBufMgr bmgrif.IBufferManager, meta *md.Table) *TableData {
 	data := &TableData{
-		Columns:     make([]col.IColumnData, 0),
 		MTBufMgr:    mtBufMgr,
 		SSTBufMgr:   sstBufMgr,
 		Meta:        meta,
+		FsMgr:       fsMgr,
 		IndexHolder: index.NewTableHolder(indexBufMgr, meta.ID),
 	}
-	for idx, colDef := range meta.Schema.ColDefs {
-		data.Columns = append(data.Columns, col.NewColumnData(data.IndexHolder, fsMgr, mtBufMgr, sstBufMgr, colDef.Type, idx))
-	}
-	runtime.SetFinalizer(data, func(o ITableData) {
-		id := o.GetID()
-		logutil2.Debugf("[GC]: TableData: %d", id)
-	})
+	data.tree.Segments = make([]iface.ISegment, 0)
+	data.tree.Helper = make(map[uint64]int)
+	data.tree.Ids = make([]uint64, 0)
+	data.OnZeroCB = data.close
+	data.Ref()
 	return data
 }
 
 type TableData struct {
-	sync.Mutex
-	RowCount    uint64
-	Columns     []col.IColumnData
+	common.RefHelper
+	tree struct {
+		sync.RWMutex
+		Segments   []iface.ISegment
+		Ids        []uint64
+		Helper     map[uint64]int
+		SegmentCnt uint32
+		RowCount   uint64
+	}
 	MTBufMgr    bmgrif.IBufferManager
 	SSTBufMgr   bmgrif.IBufferManager
 	Meta        *md.Table
 	IndexHolder *index.TableHolder
+	FsMgr       base.IManager
 }
 
-func (td *TableData) GetRowCount() uint64 {
-	return td.RowCount
+func (td *TableData) close() {
+	td.IndexHolder.Unref()
+	for _, segment := range td.tree.Segments {
+		segment.Unref()
+	}
+	// log.Infof("table %d noref", td.Meta.ID)
 }
 
 func (td *TableData) GetIndexHolder() *index.TableHolder {
 	return td.IndexHolder
 }
 
+func (td *TableData) WeakRefRoot() iface.ISegment {
+	if atomic.LoadUint32(&td.tree.SegmentCnt) == 0 {
+		return nil
+	}
+	td.tree.RLock()
+	root := td.tree.Segments[0]
+	td.tree.RUnlock()
+	return root
+}
+
+func (td *TableData) StongRefRoot() iface.ISegment {
+	if atomic.LoadUint32(&td.tree.SegmentCnt) == 0 {
+		return nil
+	}
+	td.tree.RLock()
+	root := td.tree.Segments[0]
+	td.tree.RUnlock()
+	root.Ref()
+	return root
+}
+
 func (td *TableData) GetID() uint64 {
 	return td.Meta.ID
 }
 
+func (td *TableData) GetName() string {
+	return td.Meta.Schema.Name
+}
+
 func (td *TableData) GetColTypes() []types.Type {
 	return td.Meta.Schema.Types()
-}
-
-func (td *TableData) GetCollumn(idx int) col.IColumnData {
-	if idx >= len(td.Meta.Schema.ColDefs) {
-		panic("logic error")
-	}
-	return td.Columns[idx]
-}
-
-func (td *TableData) GetCollumns() []col.IColumnData {
-	return td.Columns
 }
 
 func (td *TableData) GetColTypeSize(idx int) uint64 {
@@ -100,80 +108,299 @@ func (td *TableData) GetSSTBufMgr() bmgrif.IBufferManager {
 	return td.SSTBufMgr
 }
 
-func (td *TableData) GetSegmentCount() uint64 {
-	return td.Columns[0].SegmentCount()
+func (td *TableData) GetFsManager() base.IManager {
+	return td.FsMgr
 }
 
-func (td *TableData) UpgradeBlock(newMeta *md.Block) (blks []col.IColumnBlock) {
-	for _, column := range td.Columns {
-		blk := column.UpgradeBlock(newMeta)
-		blks = append(blks, blk)
-	}
-	return blks
+func (td *TableData) GetSegmentCount() uint32 {
+	return atomic.LoadUint32(&td.tree.SegmentCnt)
 }
 
-func (td *TableData) UpgradeSegment(segID common.ID) (segs []col.IColumnSegment) {
-	for _, column := range td.Columns {
-		seg := column.UpgradeSegment(segID)
-		segs = append(segs, seg)
+func (td *TableData) GetMeta() *md.Table {
+	return td.Meta
+}
+
+func (td *TableData) String() string {
+	td.tree.RLock()
+	defer td.tree.RUnlock()
+	s := fmt.Sprintf("<Table[%d]>(SegCnt=%d)(Refs=%d)", td.Meta.ID, td.tree.SegmentCnt, td.RefCount())
+	for _, seg := range td.tree.Segments {
+		s = fmt.Sprintf("%s\n\t%s", s, seg.String())
 	}
+
+	return s
+}
+
+func (td *TableData) WeakRefSegment(id uint64) iface.ISegment {
+	td.tree.RLock()
+	defer td.tree.RUnlock()
+	idx, ok := td.tree.Helper[id]
+	if !ok {
+		return nil
+	}
+	return td.tree.Segments[idx]
+}
+
+func (td *TableData) StrongRefSegment(id uint64) iface.ISegment {
+	td.tree.RLock()
+	defer td.tree.RUnlock()
+	idx, ok := td.tree.Helper[id]
+	if !ok {
+		return nil
+	}
+	seg := td.tree.Segments[idx]
+	seg.Ref()
+	return seg
+}
+
+func (td *TableData) WeakRefBlock(segId, blkId uint64) iface.IBlock {
+	seg := td.WeakRefSegment(segId)
+	if seg == nil {
+		return nil
+	}
+	return seg.WeakRefBlock(blkId)
+}
+
+func (td *TableData) StrongRefBlock(segId, blkId uint64) iface.IBlock {
+	seg := td.WeakRefSegment(segId)
+	if seg == nil {
+		return nil
+	}
+	return seg.StrongRefBlock(blkId)
+}
+
+func (td *TableData) RegisterSegment(meta *md.Segment) (seg iface.ISegment, err error) {
+	seg, err = NewSegment(td, meta)
+	if err != nil {
+		panic(err)
+	}
+	td.tree.Lock()
+	defer td.tree.Unlock()
+	_, ok := td.tree.Helper[meta.ID]
+	if ok {
+		return nil, errors.New("Duplicate seg")
+	}
+
+	if len(td.tree.Segments) != 0 {
+		seg.Ref()
+		td.tree.Segments[len(td.tree.Segments)-1].SetNext(seg)
+	}
+
+	td.tree.Segments = append(td.tree.Segments, seg)
+	td.tree.Ids = append(td.tree.Ids, seg.GetMeta().ID)
+	td.tree.Helper[meta.ID] = int(td.tree.SegmentCnt)
+	atomic.AddUint32(&td.tree.SegmentCnt, uint32(1))
+	seg.Ref()
+	return seg, err
+}
+
+func (td *TableData) Size(attr string) uint64 {
+	size := uint64(0)
+	segCnt := atomic.LoadUint32(&td.tree.SegmentCnt)
+	var seg iface.ISegment
+	for i := 0; i < int(segCnt); i++ {
+		td.tree.RLock()
+		seg = td.tree.Segments[i]
+		td.tree.RUnlock()
+		size += seg.Size(attr)
+	}
+	return size
+}
+
+func (td *TableData) GetSegmentedIndex() (id uint64, ok bool) {
+	ts := td.Meta.Info.GetCheckpointTime()
+	id, ok = td.Meta.CreatedIndex, true
+	if td.Meta.IsDeleted(ts) {
+		return td.Meta.DeletedIndex, true
+	}
+
+	segCnt := atomic.LoadUint32(&td.tree.SegmentCnt)
+	for i := int(segCnt) - 1; i >= 0; i-- {
+		td.tree.RLock()
+		seg := td.tree.Segments[i]
+		td.tree.RUnlock()
+		id, ok := seg.GetSegmentedIndex()
+		if ok {
+			return id, ok
+		}
+	}
+	return id, ok
+}
+
+func (td *TableData) GetReplayIndex() *md.LogIndex {
+	return td.Meta.ReplayIndex
+}
+
+func (td *TableData) SegmentIds() []uint64 {
+	ids := make([]uint64, 0, atomic.LoadUint32(&td.tree.SegmentCnt))
+	td.tree.RLock()
+	for _, seg := range td.tree.Segments {
+		ids = append(ids, seg.GetMeta().ID)
+	}
+	td.tree.RUnlock()
+	return ids
+}
+
+func (td *TableData) GetRowCount() uint64 {
+	return atomic.LoadUint64(&td.tree.RowCount)
+}
+
+func (td *TableData) AddRows(rows uint64) uint64 {
+	return atomic.AddUint64(&td.tree.RowCount, rows)
+}
+
+func (td *TableData) initReplayCtx() {
+	if td.tree.SegmentCnt == 0 {
+		return
+	}
+	var ctx *md.LogIndex
+	for segIdx := int(td.tree.SegmentCnt) - 1; segIdx >= 0; segIdx-- {
+		seg := td.tree.Segments[segIdx]
+		if ctx = seg.GetReplayIndex(); ctx != nil {
+			break
+		}
+	}
+	td.Meta.ReplayIndex = ctx
+}
+
+func (td *TableData) initRowCount() {
+	for _, seg := range td.tree.Segments {
+		td.tree.RowCount += seg.GetRowCount()
+	}
+}
+
+func (td *TableData) RegisterBlock(meta *md.Block) (blk iface.IBlock, err error) {
+	td.tree.RLock()
+	defer td.tree.RUnlock()
+	idx, ok := td.tree.Helper[meta.Segment.ID]
+	if !ok {
+		return nil, errors.New(fmt.Sprintf("seg %d not found", meta.Segment.ID))
+	}
+	seg := td.tree.Segments[idx]
+	blk, err = seg.RegisterBlock(meta)
+	return blk, err
+}
+
+func (td *TableData) UpgradeBlock(meta *md.Block) (blk iface.IBlock, err error) {
+	idx, ok := td.tree.Helper[meta.Segment.ID]
+	if !ok {
+		return nil, errors.New("seg not found")
+	}
+	seg := td.tree.Segments[idx]
+	return seg.UpgradeBlock(meta)
+}
+
+func (td *TableData) UpgradeSegment(id uint64) (seg iface.ISegment, err error) {
+	idx, ok := td.tree.Helper[id]
+	if !ok {
+		panic("logic error")
+	}
+	old := td.tree.Segments[idx]
+	if old.GetType() != base.UNSORTED_SEG {
+		panic(fmt.Sprintf("old segment %d type is %d", id, old.GetType()))
+	}
+	if old.GetMeta().ID != id {
+		panic("logic error")
+	}
+	meta, err := td.Meta.ReferenceSegment(id)
+	if err != nil {
+		return nil, err
+	}
+	upgradeSeg, err := old.CloneWithUpgrade(td, meta)
+	if err != nil {
+		panic(err)
+	}
+
+	var oldNext iface.ISegment
+	if idx != len(td.tree.Segments)-1 {
+		oldNext = old.GetNext()
+	}
+	upgradeSeg.SetNext(oldNext)
+
+	td.tree.Lock()
+	defer td.tree.Unlock()
+	td.tree.Segments[idx] = upgradeSeg
+	if idx > 0 {
+		upgradeSeg.Ref()
+		td.tree.Segments[idx-1].SetNext(upgradeSeg)
+	}
+	// old.SetNext(nil)
+	upgradeSeg.Ref()
+	old.Unref()
+	return upgradeSeg, nil
+}
+
+func MockSegments(meta *md.Table, tblData iface.ITableData) []uint64 {
+	segs := make([]uint64, 0)
+	for _, segMeta := range meta.Segments {
+		seg, err := tblData.RegisterSegment(segMeta)
+		if err != nil {
+			panic(err)
+		}
+		for _, blkMeta := range segMeta.Blocks {
+			blk, err := seg.RegisterBlock(blkMeta)
+			if err != nil {
+				panic(err)
+			}
+			blk.Unref()
+		}
+		segs = append(segs, seg.GetMeta().ID)
+		seg.Unref()
+	}
+
 	return segs
 }
 
-// Only be called at engine startup.
-func (td *TableData) AppendColSegments(colSegs []col.IColumnSegment) {
-	if len(colSegs) != len(td.Meta.Schema.ColDefs) {
-		panic("logic error")
-	}
-	for idx, column := range td.Columns {
-		if idx != column.GetColIdx() {
-			panic("logic error")
-		}
-		err := column.Append(colSegs[idx])
-		if err != nil {
-			panic(fmt.Sprintf("logic error: %s", err))
-		}
-	}
-}
-
 type Tables struct {
-	sync.RWMutex
-	Data      map[uint64]ITableData
+	Mu        *sync.RWMutex
+	Data      map[uint64]iface.ITableData
 	Ids       map[uint64]bool
-	Tombstone map[uint64]ITableData
+	Tombstone map[uint64]iface.ITableData
 }
 
-func NewTables() *Tables {
+func NewTables(mu *sync.RWMutex) *Tables {
 	return &Tables{
-		Data:      make(map[uint64]ITableData),
+		Mu:        mu,
+		Data:      make(map[uint64]iface.ITableData),
 		Ids:       make(map[uint64]bool),
-		Tombstone: make(map[uint64]ITableData),
+		Tombstone: make(map[uint64]iface.ITableData),
 	}
+}
+
+func (ts *Tables) String() string {
+	ts.Mu.RLock()
+	defer ts.Mu.RUnlock()
+	s := fmt.Sprintf("<Tables>[Cnt=%d]", len(ts.Data))
+	for _, td := range ts.Data {
+		s = fmt.Sprintf("%s\n%s", s, td.String())
+	}
+	return s
 }
 
 func (ts *Tables) TableIds() (ids map[uint64]bool) {
 	return ts.Ids
 }
 
-func (ts *Tables) DropTable(tid uint64) (err error) {
-	ts.Lock()
-	err = ts.DropTableNoLock(tid)
-	ts.Unlock()
-	return err
+func (ts *Tables) DropTable(tid uint64) (tbl iface.ITableData, err error) {
+	ts.Mu.Lock()
+	tbl, err = ts.DropTableNoLock(tid)
+	ts.Mu.Unlock()
+	return tbl, err
 }
 
-func (ts *Tables) DropTableNoLock(tid uint64) (err error) {
+func (ts *Tables) DropTableNoLock(tid uint64) (tbl iface.ITableData, err error) {
 	tbl, ok := ts.Data[tid]
 	if !ok {
-		return errors.New(fmt.Sprintf("Specified table %d not found", tid))
+		// return errors.New(fmt.Sprintf("Specified table %d not found", tid))
+		return tbl, NotExistErr
 	}
-	ts.Tombstone[tid] = tbl
+	// ts.Tombstone[tid] = tbl
 	delete(ts.Ids, tid)
 	delete(ts.Data, tid)
-	return nil
+	return tbl, nil
 }
 
-func (ts *Tables) GetTableNoLock(tid uint64) (tbl ITableData, err error) {
+func (ts *Tables) GetTableNoLock(tid uint64) (tbl iface.ITableData, err error) {
 	tbl, ok := ts.Data[tid]
 	if !ok {
 		return nil, errors.New(fmt.Sprintf("Specified table %d not found", tid))
@@ -181,21 +408,31 @@ func (ts *Tables) GetTableNoLock(tid uint64) (tbl ITableData, err error) {
 	return tbl, err
 }
 
-func (ts *Tables) GetTable(tid uint64) (tbl ITableData, err error) {
-	ts.RLock()
+func (ts *Tables) WeakRefTable(tid uint64) (tbl iface.ITableData, err error) {
+	ts.Mu.RLock()
 	tbl, err = ts.GetTableNoLock(tid)
-	ts.RUnlock()
+	ts.Mu.RUnlock()
 	return tbl, err
 }
 
-func (ts *Tables) CreateTable(tbl ITableData) (err error) {
-	ts.Lock()
+func (ts *Tables) StrongRefTable(tid uint64) (tbl iface.ITableData, err error) {
+	ts.Mu.RLock()
+	tbl, err = ts.GetTableNoLock(tid)
+	if tbl != nil {
+		tbl.Ref()
+	}
+	ts.Mu.RUnlock()
+	return tbl, err
+}
+
+func (ts *Tables) CreateTable(tbl iface.ITableData) (err error) {
+	ts.Mu.Lock()
 	err = ts.CreateTableNoLock(tbl)
-	ts.Unlock()
+	ts.Mu.Unlock()
 	return err
 }
 
-func (ts *Tables) CreateTableNoLock(tbl ITableData) (err error) {
+func (ts *Tables) CreateTableNoLock(tbl iface.ITableData) (err error) {
 	_, ok := ts.Data[tbl.GetID()]
 	if ok {
 		return errors.New(fmt.Sprintf("Dup table %d found", tbl.GetID()))
@@ -208,83 +445,35 @@ func (ts *Tables) CreateTableNoLock(tbl ITableData) (err error) {
 func (ts *Tables) Replay(fsMgr base.IManager, indexBufMgr, mtBufMgr, sstBufMgr bmgrif.IBufferManager, info *md.MetaInfo) error {
 	for _, meta := range info.Tables {
 		tbl := NewTableData(fsMgr, indexBufMgr, mtBufMgr, sstBufMgr, meta)
-		colTypes := meta.Schema.Types()
 		activeSeg := meta.GetActiveSegment()
 		for _, segMeta := range meta.Segments {
-			if activeSeg != nil {
-				if activeSeg.ID < segMeta.ID {
+			if activeSeg != nil && activeSeg.ID < segMeta.ID {
+				break
+			}
+
+			seg, err := tbl.RegisterSegment(segMeta)
+			if err != nil {
+				panic(err)
+			}
+			defer seg.Unref()
+
+			activeBlk := segMeta.GetActiveBlk()
+			for _, blkMeta := range segMeta.Blocks {
+				if activeBlk != nil && activeBlk.ID <= blkMeta.ID {
 					break
 				}
-			}
-			segType := base.UNSORTED_SEG
-			if segMeta.DataState == md.SORTED {
-				segType = base.SORTED_SEG
-			}
-			activeBlk := segMeta.GetActiveBlk()
-			for colIdx, colType := range colTypes {
-				colSeg := col.NewColumnSegment(tbl.GetIndexHolder(), fsMgr, mtBufMgr, sstBufMgr, colIdx, segMeta)
-				defer colSeg.UnRef()
-				for _, blkMeta := range segMeta.Blocks {
-					if activeBlk != nil {
-						if activeBlk.ID <= blkMeta.ID {
-							break
-						}
-					}
-					blkId := *blkMeta.AsCommonID()
-					bufMgr := mtBufMgr
-					if segType == base.SORTED_SEG {
-						bufMgr = sstBufMgr
-					} else if blkMeta.DataState == md.FULL {
-						bufMgr = sstBufMgr
-					}
-					// TODO: strblk
-					// Only stdblk now
-					colBlk := col.NewStdColumnBlock(colSeg.Ref(), blkMeta)
-					colSeg.Append(colBlk.Ref())
-					defer colBlk.UnRef()
-					colPart := col.NewColumnPart(colSeg.GetFsManager(), bufMgr, colBlk.Ref(), blkId, info.Conf.BlockMaxRows*uint64(colType.Size))
-					if colPart == nil {
-						return errors.New(fmt.Sprintf("data replay error"))
-					}
-					// TODO: How to handle more than 1 empty blk or segment
-					// if blkMeta.DataState == md.EMPTY {
-					// 	break
-					// }
+				blk, err := seg.RegisterBlock(blkMeta)
+				if err != nil {
+					panic(err)
 				}
-				if err := tbl.GetCollumn(colIdx).Append(colSeg.Ref()); err != nil {
-					return err
-				}
+				defer blk.Unref()
 			}
 		}
+		tbl.initRowCount()
+		tbl.initReplayCtx()
 		if err := ts.CreateTable(tbl); err != nil {
 			return err
 		}
 	}
 	return nil
-}
-
-func MockSegment(indexHolder *index.TableHolder, fsMgr base.IManager, mtBufMgr, sstBufMgr bmgrif.IBufferManager, colIdx int, meta *md.Segment) col.IColumnSegment {
-	seg := col.NewColumnSegment(indexHolder, fsMgr, mtBufMgr, sstBufMgr, colIdx, meta)
-	for _, blkMeta := range meta.Blocks {
-		blk, err := seg.RegisterBlock(blkMeta)
-		if err != nil {
-			panic(err)
-		}
-		blk.UnRef()
-	}
-	return seg
-}
-
-func MockSegments(fsMgr base.IManager, mtBufMgr, sstBufMgr bmgrif.IBufferManager, meta *md.Table, tblData ITableData) []common.ID {
-	var segIDs []common.ID
-	for _, segMeta := range meta.Segments {
-		var colSegs []col.IColumnSegment
-		for colIdx, _ := range segMeta.Table.Schema.ColDefs {
-			colSeg := MockSegment(tblData.GetIndexHolder(), fsMgr, mtBufMgr, sstBufMgr, colIdx, segMeta)
-			colSegs = append(colSegs, colSeg)
-		}
-		tblData.AppendColSegments(colSegs)
-		segIDs = append(segIDs, *segMeta.AsCommonID())
-	}
-	return segIDs
 }
