+++ pkg/vm/engine/aoe/storage/layout/table/v1/handle/blk.go
@@ -1,77 +1,36 @@
 package handle
 
 import (
-	"fmt"
-	logutil2 "matrixone/pkg/logutil"
-	"matrixone/pkg/vm/engine/aoe/storage/common"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/index"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/col"
-	"matrixone/pkg/vm/engine/aoe/storage/mock/type/chunk"
-	"sync"
+	"matrixone/pkg/vm/engine/aoe/storage/dbi"
 )
 
 var (
-	// allocTimes    = 0
-	blkHandlePool = sync.Pool{
-		New: func() interface{} {
-			// allocTimes++
-			// log.Infof("Alloc blk handle: %d", allocTimes)
-			h := new(BlockHandle)
-			h.Cols = make([]col.IColumnBlock, 0)
-			// h.Cursors = make([]col.ScanCursor, 0)
-			return h
-		},
-	}
+	_ dbi.IBlock = (*Block)(nil)
 )
 
-type BlockHandle struct {
-	ID          common.ID
-	Cols        []col.IColumnBlock
-	IndexHolder *index.BlockHolder
+type Block struct {
+	Host *Segment
+	Id   uint64
 }
 
-func (bh *BlockHandle) GetID() *common.ID {
-	return &bh.ID
+func (blk *Block) Prefetch() dbi.IBatchReader {
+	realBlk := blk.Host.Data.StrongRefBlock(blk.Id)
+	defer realBlk.Unref()
+	return realBlk.GetBatch(blk.Host.Attr)
 }
 
-func (bh *BlockHandle) GetColumn(idx int) col.IColumnBlock {
-	if idx < 0 || idx >= len(bh.Cols) {
-		panic(fmt.Sprintf("Specified idx %d is out of scope", idx))
-	}
-	return bh.Cols[idx]
+func (blk *Block) GetID() uint64 {
+	return blk.Id
 }
 
-func (bh *BlockHandle) GetIndexHolder() *index.BlockHolder {
-	return bh.IndexHolder
+func (blk *Block) GetSegmentID() uint64 {
+	realBlk := blk.Host.Data.StrongRefBlock(blk.Id)
+	defer realBlk.Unref()
+	return realBlk.GetMeta().Segment.ID
 }
 
-func (bh *BlockHandle) Close() error {
-	if bhh := bh; bhh != nil {
-		for _, col := range bhh.Cols {
-			col.UnRef()
-		}
-		bhh.Cols = bhh.Cols[:0]
-		// TODO
-		// blkHandlePool.Put(bhh)
-		bh = nil
-	}
-	return nil
-}
-
-func (bh *BlockHandle) InitScanCursor() []col.ScanCursor {
-	cursors := make([]col.ScanCursor, len(bh.Cols))
-	for idx, colBlk := range bh.Cols {
-		colBlk.InitScanCursor(&cursors[idx])
-		err := cursors[idx].Init()
-		if err != nil {
-			logutil2.Error(fmt.Sprintf("logic error: %s", err))
-			panic(fmt.Sprintf("logic error: %s", err))
-		}
-	}
-	return cursors
-}
-
-func (bh *BlockHandle) Fetch() *chunk.Chunk {
-	// TODO
-	return nil
+func (blk *Block) GetTableID() uint64 {
+	realBlk := blk.Host.Data.StrongRefBlock(blk.Id)
+	defer realBlk.Unref()
+	return realBlk.GetMeta().Segment.Table.ID
 }
