+++ pkg/vm/engine/aoe/storage/layout/table/v1/handle/seglinkit.go
@@ -1,50 +1,63 @@
 package handle
 
 import (
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/col"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/handle/base"
-	// log "github.com/sirupsen/logrus"
+	"matrixone/pkg/vm/engine/aoe/storage/dbi"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/iface"
 )
 
 var (
-	_ base.ISegmentIterator = (*SegmentIt)(nil)
+	_ dbi.ISegmentIt = (*SegmentLinkIt)(nil)
 )
 
 type SegmentLinkIt struct {
-	Cols []col.IColumnSegment
+	OnCloseCB CloseSegmentItCB
+	Snapshot  *Snapshot
+	Cursor    iface.ISegment
 }
 
-func (it *SegmentLinkIt) Valid() bool {
-	if it == nil {
-		return false
+func NewSegmentLinkIt(ss *Snapshot) dbi.ISegmentIt {
+	if ss == nil {
+		return &SegmentLinkIt{}
 	}
-	if len(it.Cols) == 0 {
-		return false
+	it := &SegmentLinkIt{
+		Snapshot:  ss,
+		OnCloseCB: ss.removeIt,
 	}
-	return true
+	it.Cursor = ss.TableData.StongRefRoot()
+	return it
 }
 
 func (it *SegmentLinkIt) Close() error {
+	if it.Cursor != nil {
+		it.Cursor.Unref()
+		it.Cursor = nil
+	}
+	if it.OnCloseCB != nil {
+		it.OnCloseCB(it)
+		it.OnCloseCB = nil
+	}
 	return nil
 }
 
 func (it *SegmentLinkIt) Next() {
-	newCols := make([]col.IColumnSegment, 0)
-	for _, colSeg := range it.Cols {
-		newCol := colSeg.GetNext()
-		if newCol == nil {
-			it.Cols = nil
-			return
-		}
-		newCols = append(newCols, newCol)
+	if it.Cursor != nil {
+		cursor := it.Cursor
+		it.Cursor = it.Cursor.GetNext()
+		cursor.Unref()
 	}
-	it.Cols = newCols
 }
 
-func (it *SegmentLinkIt) GetSegmentHandle() base.ISegmentHandle {
-	h := &SegmentHandle{
-		ID:   it.Cols[0].GetID(),
-		Cols: it.Cols,
+func (it *SegmentLinkIt) Valid() bool {
+	if it.Cursor != nil {
+		return true
 	}
-	return h
+	return false
+}
+
+func (it *SegmentLinkIt) GetHandle() dbi.ISegment {
+	seg := &Segment{
+		Data: it.Snapshot.TableData.WeakRefSegment(it.Cursor.GetMeta().ID),
+		Attr: it.Snapshot.Attr,
+	}
+	return seg
 }
