+++ pkg/vm/engine/aoe/storage/layout/table/v1/handle/ss.go
@@ -2,7 +2,7 @@ package handle
 
 import (
 	"matrixone/pkg/vm/engine/aoe/storage/dbi"
-	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v2/iface"
+	"matrixone/pkg/vm/engine/aoe/storage/layout/table/v1/iface"
 	"sync"
 	"sync/atomic"
 )
