+++ pkg/vm/engine/aoe/storage/layout/table/v1/wrapper/batch.go
