+++ pkg/vm/engine/aoe/storage/layout/table/v1/wrapper/vector.go
